Add nvm configuration to zsh environment
If using Node Version Manager (nvm), this change adds a new configuration file
to the zsh environment. It checks if the nvm installation exists in the user's
home directory, and if so, it exports the NVM_DIR environment variable and
sources the nvm.sh script.

This allows nvm to manage and switch between different Node.js versions
seamlessly in the zsh shell environment.