Install fnm and add it to sheldon plugins
fnm (Fast Node Manager) is a utility to manage multiple Node.js versions.
This commit adds fnm to the list of Rust packages installed by the setup script.
It also adds fnm to the list of sheldon plugins and configures it to load automatically in zsh.
Additionally, it moves some environment configuration files to the .zshrc.d directory,
and adds a global .stowignore file to exclude .DS_Store files from being tracked.